foodgram/backend/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
pillow==11.0.0
shortuuid
transliterate
django-filter==23.5
psycopg2-binary
